docs: update example project to new loader-first strategy
