--- example/src/ProtectedRoute.tsx
@@ -1,3 +1,0 @@
-export default function ProtectedRoute() {
-  return <div>This route is protected by `withAuthenticationRequired`</div>
-}
